[MAILDIR] Implement moving message between Maildir sources based on label.
To use this functionality, in config.yaml, you need a new option
:maildir_labels:

    :maildir_labels:
      :stanford: [[:inbox, 4], [null, 6]]

The value of this option is a dictionary of "accounts" to lists of
precedences.  Read it as follows:

    For messages in source 4 or source 6 (consult sources.yaml),
    if the message has the :inbox tag, move it to source 4, otherwise
    move it to source 6.

This will automatically start working for any new mail you change the
labels of.  In order to apply this to old mail, you need to run
sup-sync-back-maildir.  If you're going to move a lot of mail, you
probably want the companion patch "Write moves between Maildirs to
'mv' directory, for OfflineIMAP" which enables efficient same-server
IMAP moving.

Signed-off-by: Edward Z. Yang <ezyang@mit.edu>

Conflicts:
	lib/sup/maildir.rb
	lib/sup/message.rb